Aquinas, removed compiler warning
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@8989 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Aquinas/Aquinas.cpp
@@ -204,7 +204,7 @@ int AqController::OnValveOnOff(MM::PropertyBase* pProp, MM::ActionType eAct, lon
    if (eAct == MM::BeforeGet)
    {
       unsigned char t = valveState_;
-      for (unsigned int i = 0; i < valveNr; i++)
+      for (unsigned int i = 0; i < (unsigned int) valveNr; i++)
          t = t >> 1;
       long val = 0;
       if (t & 1)
